requirements.txt
allure-pytest==2.15.3
allure-python-commons==2.15.3
annotated-types==0.7.0
attrs==26.1.0
autoflake==2.3.3
black==26.3.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
execnet==2.1.2
flake8==7.3.0
h11==0.16.0
idna==3.11
iniconfig==2.3.0
logging==0.4.9.6
mccabe==0.7.0
mypy_extensions==1.1.0
outcome==1.3.0.post0
packaging==26.1
pathspec==1.0.4
platformdirs==4.9.6
pluggy==1.6.0
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
pyflakes==3.4.0
Pygments==2.20.0
PySocks==1.7.1
pytest==9.0.3
pytest-xdist==3.8.0
pytokens==0.4.1
requests==2.33.1
selenium==4.43.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.33.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websocket-client==1.9.0
wsproto==1.3.2
